requirements.txt
flask
pandas
azure-ai-textanalytics
matplotlib
wordcloud
seaborn